Implémentation des fonctionnalité du dashboard Admin (Gestion des Boutiques )

+++ database/migrations/2025_04_23_162537_add_is_blocked_to_users_table.php
@@ -11,9 +11,15 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->boolean('is_blocked')->default(false)->after('is_active');
-        });
+       /** Schema::table('users', function (Blueprint $table) {
+           * $table->boolean('is_blocked')->default(false)->after('is_active');
+        });*/
+        if (!Schema::hasColumn('users', 'is_blocked')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->boolean('is_blocked')->default(false)->after('is_active');
+            });
+        }
+
     }
 
     /**
@@ -21,8 +27,15 @@ return new class extends Migration
      */
     public function down(): void
     {
+       /* Schema::table('users', function (Blueprint $table) {
+            $table->dropColumn('is_blocked');
+        });*/
+
+
+    if (Schema::hasColumn('users', 'is_blocked')) {
         Schema::table('users', function (Blueprint $table) {
             $table->dropColumn('is_blocked');
         });
+    }
     }
 };
